Fix bad Forme setting due to integer conversion

+++ pk3DS/Subforms/Gen7/SMWE.cs
@@ -642,7 +642,7 @@ namespace pk3DS
                                     }
                                     encounter.Species = (uint)species;
                                 }
-                                encounter.Forme = (uint)GetRandomForme((int)encounter.Species);
+                                encounter.Forme = GetRandomForme((int)encounter.Species);
                             }
                         }
                     }
@@ -655,16 +655,16 @@ namespace pk3DS
             Util.Alert("Randomized!");
         }
 
-        public int GetRandomForme(int species)
+        public uint GetRandomForme(int species)
         {
             if (Main.SpeciesStat[species].FormeCount > 1)
             {
                 if (Legal.Mega_ORAS.Contains((ushort)species))
                     if (CHK_MegaForm.Checked)
-                        return (int) rnd32()%Main.SpeciesStat[species].FormeCount; // Slot-Random
+                        return (uint) (rnd32()%Main.SpeciesStat[species].FormeCount); // Slot-Random
                     else
                         return 0;
-                return (int)rnd32() % Main.SpeciesStat[species].FormeCount; // Slot-Random
+                return (uint)(rnd32() % Main.SpeciesStat[species].FormeCount); // Slot-Random
 
             }
             return 0;
